Move loose parse int to its own file
