+++ source/features/deep-reblame.tsx
@@ -10,7 +10,8 @@ import features from '.';
 import * as api from '../github-helpers/api';
 import GitHubURL from '../github-helpers/github-url';
 import LoadingIcon from '../github-helpers/icon-loading';
-import {getRepoGQL, looseParseInt} from '../github-helpers';
+import {getRepoGQL} from '../github-helpers';
+import looseParseInt from '../helpers/loose-parse-int';
 
 const getPullRequestBlameCommit = mem(async (commit: string, prNumber: number, currentFilename: string): Promise<string> => {
 	const {repository} = await api.v4(`
