+++ source/helpers/loose-parse-int.ts
@@ -1,0 +1,3 @@
+export default function looseParseInt(text: string): number {
+	return Number(text.replace(/\D+/g, ''));
+}
